+++ lib/lib_ssl/bearssl-esp8266/src/t_bearssl_x509.h
@@ -1032,7 +1032,7 @@ typedef struct {
  * \param append_in       issuer DN receiver callback (or `0`).
  * \param append_in_ctx   context for the issuer DN receiver callback.
  */
-void br_x509_decoder_init(br_x509_decoder_context *ctx,
+void br_x509_decoder_init_esp8266(br_x509_decoder_context *ctx,
 	void (*append_dn)(void *ctx, const void *buf, size_t len),
 	void *append_dn_ctx,
 	void (*append_in)(void *ctx, const void *buf, size_t len),